cleaned up some folder structure. Added more to login page.
