+++ client/src/components/Nav/index.js
@@ -4,7 +4,7 @@ function Nav() {
   return (
     <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
       <a className="navbar-brand" href="/">
-        Stem 🌱
+        Stem <span role="img" aria-label="Logo">🌱</span>
       </a>
     </nav>
   );
