--- client/src/pages/Login.js
@@ -1,22 +1,0 @@
-import React, { Component } from "react";
-
-class Login extends Component {
-
-    render() {
-        return (
-            <div>
-                <div className="container text-center">
-                    <div className="jumbotron">
-                        <h1 className="display-4">Stem</h1>
-                        <p className="lead">An app for easy stalking</p>
-                        <hr className="my-4" />
-                        <p>It uses utility classes for typography and spacing to space content out within the larger container.</p>
-                        <a className="btn btn-primary btn-lg" href="#" role="button">Learn more</a>
-                    </div>
-                </div>
-            </div>
-        )
-    }
-}
-
-export default Login;
